removed suo file.  Greatly improved overall stability (espectially for BFGS).  Changed the position weighting function.  Tweaked the matlab example.

+++ include/nanoflann/nanoflann.hpp
@@ -312,10 +312,10 @@ namespace nanoflann
 
 			/* Process 4 items with each loop for efficiency. */
 			while (a < lastgroup) {
-				const DistanceType diff0 = a[0] - data_source.kdtree_get_pt(b_idx,d++);
-				const DistanceType diff1 = a[1] - data_source.kdtree_get_pt(b_idx,d++);
-				const DistanceType diff2 = a[2] - data_source.kdtree_get_pt(b_idx,d++);
-				const DistanceType diff3 = a[3] - data_source.kdtree_get_pt(b_idx,d++);
+				const DistanceType diff0 = (int)(a[0] - data_source.kdtree_get_pt(b_idx,d++));
+				const DistanceType diff1 = (int)(a[1] - data_source.kdtree_get_pt(b_idx,d++));
+				const DistanceType diff2 = (int)(a[2] - data_source.kdtree_get_pt(b_idx,d++));
+				const DistanceType diff3 = (int)(a[3] - data_source.kdtree_get_pt(b_idx,d++));
 				result += diff0 * diff0 + diff1 * diff1 + diff2 * diff2 + diff3 * diff3;
 				a += 4;
 				if ((worst_dist>0)&&(result>worst_dist)) {
